pyproject.toml
[tool.poetry]
name = "ai_product_pilot"
version = "0.1.0"
description = "Un agent d'analyse de feedback utilisateur avec LangGraph et Supabase"
authors = ["Charlie Quillard <charlie.quillard@gmail.com>"]
packages = [{ include = "ai_product_pilot" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
uvicorn = "^0.27.1"
pydantic = "^2.6.3"
langchain = "^0.3.24"
langchain-openai = "^0.3.15"
langchain-community = "^0.3.23"
langgraph = "^0.4.1"
supabase = "^2.3.4"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"
langsmith = "^0.3.38"
pydantic-settings = "^2.2.1"
celery = "^5.3.6"
sqlmodel = "^0.0.14"
pytest = "^8.0.2"
pytest-asyncio = "^0.23.5"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
